Объединила модули по открытию и закрытию формы загрузки фотографии пользователя, изменению маштаба фото, наложению эффектов на изображение, а также валидацию полей вводахэштегов и комментариев в один form-upload-img
