+++ js/main.js
@@ -1,16 +1,16 @@
 import './function.js';
 import './utils.js';
-import './pattern-photo.js';
+// import './pattern-photo.js';
 import './rendering-photo.js';
 import './fullsize-photo.js';
-import './fullsize-photo-close.js';
-import './fullsize-photo-open.js';
-import './form-upload-img-open.js';
-import './form-upload-img-close.js';
-import './form-validation.js';
-import './form-change-size-img.js';
-import './form-effects-img.js';
-import './remove-event-listeners.js';
+// import './fullsize-photo-close.js';
+// import './fullsize-photo-open.js';
+import './form-upload-img.js';
+// import './form-upload-img-close.js';
+// import './form-validation.js';
+// import './form-change-size-img.js';
+// import './form-effects-img.js';
+// import './remove-event-listeners.js';
 import './server.js';
 import './popups.js';
 import './popup-close.js';
